+++ query.d.ts
@@ -1,0 +1,8 @@
+/* Query Interfaces */
+/* https://rdf.js.org/query-spec/ */
+
+export * from './query/common';
+export * from './query/filterable';
+export * from './query/queryable';
+
+